+++ vscode.tests/tests.js
@@ -1,26 +1,24 @@
 const assert = require('assert')
 const {window, workspace} = require('vscode')
-const getSettings = require('../vscode/compiled/Settings').default
+const {getCoreSettings} = require('../vscode/compiled/Settings')
 
 const testSettings = async (language, expected) => {
     const doc = await workspace.openTextDocument({language, content: ""})
     const editor = await window.showTextDocument(doc)
-    const actual = getSettings(editor)
+    const actual = getCoreSettings(editor, cs => cs[0])
     assert.deepStrictEqual(actual, expected)
 }
 
 const settingsData = {
     plaintext: {
-        column: undefined,
-        columns: [45, 90],
+        column: 45,
         doubleSentenceSpacing: true,
         wholeComment: false,
         reformat: true,
         tabWidth: 2,
     },
     markdown: {
-        column: undefined,
-        columns: [60],
+        column: 60,
         doubleSentenceSpacing: false,
         wholeComment: true,
         reformat: false,