+++ CHANGELOG.md
@@ -3,6 +3,8 @@ See also [https://github.com/stkb/vscode-rewrap/releases](https://github.com/stk
 
 ## Unreleased
 
+- Make `autoWrap.enabled` a per-language setting; auto-wrap toggle command a
+  per-document override. (#149)
 - Add Clojure, Common Lisp, Emacs Lisp, Scheme, and J (#246, #240)
 - Fix per-language settings sometimes not applying.
 - Declare as a UI extension so it doesn't have to be installed in the remote